hongs-magpie-web/web/WEB-INF/bin/doc-parser/requirements.txt
Flask
python-dotenv
markitdown[all]
markitdown-ocr
openai